+++ core/datetimetools.py
