+++ data/MIDRC Open A1 and R1 - cumulative by batch.xlsx
