+++ data/MIDRC Open A1 and R1 COVIDpos only - cumulative by batch.xlsx
