+++ jsdconfig-example.yaml
@@ -3,24 +3,24 @@ data sources:
   - name: MIDRC
     description: MIDRC Excel File
     data type: file
-    filename: MIDRC Open A1 and R1 - cumulative by batch.xlsx
+    filename: data/MIDRC Open A1 and R1 - cumulative by batch.xlsx
     remove column name text: [(CUSUM)]
 
   - name: CDC
     description: CDC Excel File
     data type: file
-    filename: CDC_COVIDpos - cumulative by month.xlsx
+    filename: data/CDC_COVIDpos - cumulative by month.xlsx
 
   - name: Census
     description: Census Excel File
     data type: file
-    filename: Census_all.xlsx
+    filename: data/Census_all.xlsx
     date: '2020-01-01' # The census file does not have a date column, so we specify the date here
 
   - name: MIDRC COVID+
     description: MIDRC COVID+ Excel File
     data type: file
-    filename: MIDRC Open A1 and R1 COVIDpos only - cumulative by batch.xlsx
+    filename: data/MIDRC Open A1 and R1 COVIDpos only - cumulative by batch.xlsx
     remove column name text: [(CUSUM)]
 
 # TODO: The following should be moved into QSettings for modifications within the GUI
